[OOPS]  Added better handling of different version

+++ gists/centos_python.sh
@@ -4,7 +4,9 @@ yum -y install gcc openssl-devel bzip2-devel wget
 mkdir -p /u/downloads
 url="https://www.python.org/ftp/python/$version/Python-$version.tgz"
 filename="/u/downloads/python_$version.tgz"
-wget $url -O $filename
+if [[ ! -f $filename ]] ; then
+    wget $url -O $filename
+]]
 mkdir -p /u/python_$version
 cd /u/python_$version
 tar xzf $filename --strip-components=1
@@ -12,6 +14,7 @@ tar xzf $filename --strip-components=1
 make altinstall
 cd /u/downloads
 wget https://bootstrap.pypa.io/get-pip.py
-python2.7 get-pip.py
+suffix=${version%.*}
+"python${suffix}" get-pip.py
